Wonder movies play again:
Some cleanings and compiler warning fixes:
modified trunk/ctp2_code/ai/ctpai.cpp
modified trunk/ctp2_code/ai/diplomacy/proposalresponseevent.cpp
modified trunk/ctp2_code/ai/mapanalysis/mapanalysis.cpp
modified trunk/ctp2_code/ai/strategy/goals/Goal.cpp
modified trunk/ctp2_code/gs/gameobj/Player.cpp
modified trunk/ctp2_code/ui/aui_common/aui_moviebutton.cpp

Compiling fixes:
modified trunk/ctp2_code/ui/aui_common/aui_music.cpp
modified trunk/ctp2_code/ui/aui_common/aui_textfield.cpp

Manual double buffering is now done with DirectX components only, and fixed the wonder movies:
modified trunk/ctp2_code/ui/aui_common/aui_blitter.cpp
modified trunk/ctp2_code/ui/aui_common/aui_mouse.cpp
modified trunk/ctp2_code/ui/aui_common/aui_ui.cpp
modified trunk/ctp2_code/ui/aui_common/aui_ui.h
modified trunk/ctp2_code/ui/aui_ctp2/c3ui.cpp
modified trunk/ctp2_code/ui/aui_ctp2/directvideo.cpp
modified trunk/ctp2_code/ui/aui_directx/aui_directblitter.cpp
modified trunk/ctp2_code/ui/aui_directx/aui_directmouse.cpp
modified trunk/ctp2_code/ui/aui_directx/aui_directmovie.cpp
modified trunk/ctp2_code/ui/aui_directx/aui_directsurface.cpp
modified trunk/ctp2_code/ui/aui_directx/aui_directui.cpp

git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@1090 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/ai/ctpai.cpp
@@ -1552,7 +1552,7 @@ void CtpAi::MakeRoomForNewUnits(const PLAYER_INDEX playerId)
 						                           false
 						                          );
 
-						defense_strength -= city.GetDefendersBonus() * static_cast<double>(defense_count);
+						defense_strength -= static_cast<float>(city.GetDefendersBonus() * static_cast<double>(defense_count));
 
 						double prev_city_defense = city->GetCityData()->GetCurrentGarrisonStrength();
 						city->GetCityData()->SetCurrentGarrisonStrength( prev_city_defense + defense_strength );
@@ -1956,7 +1956,7 @@ void CtpAi::ComputeCityGarrisons(const PLAYER_INDEX playerId )
 		                      false
 		                     );
 
-		defense_strength += city.GetDefendersBonus() * static_cast<double>(defense_count);
+		defense_strength += static_cast<float>(city.GetDefendersBonus() * static_cast<double>(defense_count));
 
 		double prev_city_defense = city->GetCityData()->GetCurrentGarrisonStrength();
 		city->GetCityData()->SetCurrentGarrisonStrength( prev_city_defense + defense_strength );
